Improve module layout
